Split the omnibox and site search out to a follow-up PR
