--- apps/studio/src/lib/wordpress-rest-api.ts
@@ -1,32 +1,0 @@
-import {
-	createJsonResponse,
-	fetchSiteRest as fetchSiteRestShared,
-} from '@studio/common/lib/wordpress-rest';
-import { SiteServer } from 'src/site-server';
-import type { SiteRestRequest, SiteRestResponse } from '@studio/common/types/wordpress-rest';
-import type { IpcMainInvokeEvent } from 'electron';
-
-export async function fetchSiteRest(
-	_event: IpcMainInvokeEvent,
-	siteId: string,
-	request: SiteRestRequest
-): Promise< SiteRestResponse > {
-	const server = SiteServer.get( siteId );
-	if ( ! server ) {
-		return createJsonResponse( 404, 'studio_site_not_found', `Site ${ siteId } not found.` );
-	}
-
-	const baseUrl =
-		server.details.port > 0
-			? `http://127.0.0.1:${ server.details.port }`
-			: server.server.url.replace( /\/+$/, '' );
-
-	return fetchSiteRestShared(
-		{
-			siteId,
-			running: server.details.running,
-			baseUrl,
-		},
-		request
-	);
-}
